Got rid of string zeros and fixed Lab 14 double issue

+++ Week14/Week14Inputs.cpp
@@ -123,7 +123,7 @@ int main()
         cout << "Height: ";
         cin >> height;
 
-        int weight;
+        double weight;
         cout << "Weight: ";
         cin >> weight;
